Only super user can view all users and stores, also added sidebar

+++ seller/delete_users.php
@@ -3,9 +3,10 @@ include("../connection/connect.php");
 error_reporting(0);
 session_start();
 
-
-// sending query
-mysqli_query($db,"DELETE FROM users WHERE u_id = '".$_GET['user_del']."'");
-header("location:allusers.php");  
-
+if (isset($_SESSION["adm_co"]) && ($_SESSION["adm_co"] == "SUPA"))
+{
+    // sending query
+    mysqli_query($db,"DELETE FROM users WHERE u_id = '".$_GET['user_del']."'");
+    header("location:allusers.php");  
+}
 ?>
